pyproject.toml
[project]
name = "vender-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Caio A. Soares",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "sqlalchemy[asyncio] (>=2.0.40,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "psycopg[binary] (>=3.2.7,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "opentelemetry-distro (>=0.53b1,<0.54)",
    "opentelemetry-instrumentation-fastapi (>=0.53b1,<0.54)",
    "opentelemetry-exporter-otlp (>=1.32.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
taskipy = "^1.14.1"
ruff = "^0.11.7"
pytest-asyncio = "^0.26.0"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT','FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev vender_api/app.py --port 8001'
pre_test = 'task lint'
test = 'pytest -s -x --cov=vender_api -vv'
post_test = 'coverage html'
rm_build = 'docker rm -f vender-api:latest'
build = 'docker build -t vender-api:latest .'